pyproject.toml
[tool.poetry]
name = "ohte"
version = "0.1.0"
description = ""
authors = ["Topias Muukkonen <topias.muukkonen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^2.2.0"
requests = "^2.31.0"
bidict = "^0.22.1"
sqlitedict = "^2.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
coverage = "^7.3.2"
pylint = "^3.0.2"
autopep8 = "^2.0.4"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
